updated user dialog
